ipq40xx: add initial sony-mcu driver for Sony NCP-HG100
This commit adds initial support for the MCU on NCP-HG100, as a MFD
(Multi-Function Device).

This MCU (Micro Controller Unit) is a Nuvoton MINI54FDE and customized
for Sony NCP-HG100 by the firmware in the MCU.

The MCU controlls the following devices:

- RGB LED tape surrounds the case of NCP-HG100
- Fan
- Temperature Sensors (2x)

Currently, RGB LED part is supported in this driver.

following properties are required:

- boot-gpio: a GPIO pin to select bootmode of the MCU
- reset-gpio: a GPIO pin to reset the MCU

DeviceTree:
-----------
mcu@15 {
	compatible = "sony,mcu";
	reg = <0x15>;

	boot-gpio = <&tlmm 38 GPIO_ACTIVE_HIGH>;
	reset-gpio = <&tlmm 5 GPIO_ACTIVE_HIGH>;
};
-----------

This driver is based on rb4xx-cpld driver in ath79 target and mcu-i2c
driver in the GPL tar provided by Sony.

Signed-off-by: INAGAKI Hiroshi <musashino.open@gmail.com>