ipq40xx: add initial leds-sony-mcu driver for Sony NCP-HG100
This commit adds initial support for RGB LED connected to the MCU on
Sony NCP-HG100.

The is the basic LED function support.

Only Red/Green/Blue color should be registered, so "color" property in
each LED node in dts is necessary.

DeviceTree:
-----------
leds {
	compatible = "sony,mcu-leds";

	red {
		label = "red:rgb";
		color = <LED_COLOR_ID_RED>;
		function = LED_FUNCTION_POWER;
		function-enumerator = <1>;
	};

	green {
		label = "green:rgb";
		color = <LED_COLOR_ID_GREEN>;
		function = LED_FUNCTION_POWER;
		function-enumerator = <2>;
	};

	blue {
		label = "blue:rgb";
		color = <LED_COLOR_ID_BLUE>;
		function = LED_FUNCTION_POWER;
		function-enumerator = <3>;
	};
};
-----------

This driver is based on leds-pwm driver in the Linux Kernel.

Signed-off-by: INAGAKI Hiroshi <musashino.open@gmail.com>